automation_pipeline/vertex_pipeline/requirements.txt
google-cloud-bigquery
google-cloud-bigquery-storage
google-cloud-storage
pandas
pyarrow
db-dtypes
numpy
joblib
scikit-learn
pytz
networkx
tqdm
matplotlib
seaborn
scipy
pandas-gbq
